changed config to yaml, restructured directories, minor script updates
