+++ cfg/benchmark.yaml
@@ -1,0 +1,39 @@
+output_dir: /nfs/diskstation/projects/dex-net/segmentation/results_test/sd_mask_test
+save_conf_name: config.yaml
+
+model:
+  path: /nfs/diskstation/projects/dex-net/segmentation/datasets/no_noise_sim_dataset/model_weights/res50_update20180620T1706/mask_rcnn_res50_update_0150.h5
+  settings:
+    mean_pixel: [119,119,119]
+    image_shape: [512,512,3]
+    backbone: resnet50
+    num_classes: 2
+    gpu_count: 1
+    images_per_gpu: 1
+
+test:
+  path: /nfs/diskstation/projects/dex-net/segmentation/datasets/real_test_cases_06_07_18/phoxi/images
+  indices: test_indices.npy
+  img_type: depth_ims
+
+mask:                                                                                                                                                              
+  remove_bin_pixels: 1              # Set to True if we want to discard predictions based on if they have bin pixels
+  bin_masks: segmasks_filled        # Directory containing bin masks                                                                                                                                                                                                                         
+  overlap_thresh: 0.5               # Overlap threshold
+
+# Visualization settings
+vis:
+  # Prediction vis settings
+  predictions: 1
+  show_bbox_pred: 0
+  show_scores_pred: 0
+  show_class_pred: 0
+
+  # GT vis settings
+  ground_truth: 0
+  show_bbox_gt: 1
+  show_scores_gt: 1
+  show_class_gt: 1
+
+  # Saurabh vis settings
+  s_bench: 0
