+++ cfg/resize.yaml
@@ -1,0 +1,5 @@
+dataset_path: ./only_noise_dataset  # base path of the directory of the dataset
+img_dir: depth_ims_1e-4 # name of the image directory within the dataset folder
+mask_dir: modal_segmasks_unscaled # name of the mask directory within the dataset folder
+img_out_dir: depth_ims # name of the output directory for images (within the dataset folder)
+mask_out_dir: modal_segmasks # name of the output directory for masks (within the dataset folder)
