+++ cfg/train.yaml
@@ -1,0 +1,24 @@
+model:
+  log_path: /nfs/diskstation/projects/dex-net/segmentation/models/
+  weights: new # new, last, coco, imagenet or custom path
+  epochs: 150      # number of epochs to train for
+  settings:
+    name: new
+    mean_pixel: [119,119,119]
+    image_shape: [512,512,3]
+    backbone: resnet50
+    num_classes: 2
+    gpu_count: 1
+    images_per_gpu: 4
+    learning_rate: 0.01
+
+dataset:
+  path: /nfs/diskstation/projects/dex-net/segmentation/datasets/wisdom/wisdom-sim
+  train_indices: train_indices.npy
+  val_indices: test_indices.npy
+  img_type: depth_ims
+
+# name of the config file to save (within the dataset folder)
+# note that this will be overwritten between multiple runs so
+# you should name this uniquely for each run.
+save_conf_name: config_1e-2_e150_res50_resume.yaml
