--- noise/config.ini
@@ -1,151 +1,0 @@
-[GENERAL]
-# set to activate a section of the configuration file
-task = "benchmark"
-
-[AUGMENT]
-# location of all the images to be augmented
-img_dir = "/nfs/diskstation/projects/dex-net/segmentation/datasets/segmasks_04_13_18/depth_ims"
-
-# output location of the augmented images
-out_dir = "./only_inpainting_dataset/depth_ims_1e-4"
-
-# use noise injection
-with_noise = True
-
-# inpaint images
-with_inpainting = False
-
-# level of noise injection (given that MAX_VAL of a pixel is 1.0)
-noise_level = 0.01
-
-# number of images to process (image_0009999.png will be the last if this is 10000)
-num_imgs = 10000
-
-
-[RESIZE]
-# base path of the directory of the dataset
-base_path = "./only_noise_dataset"
-
-# name of the image directory within the dataset folder
-img_dir = "depth_ims_1e-4"
-
-# name of the mask directory within the dataset folder
-mask_dir = "modal_segmasks_unscaled"
-
-# name of the output directory for images (within the dataset folder)
-img_out_dir = "depth_ims"
-
-# name of the output directory for masks (within the dataset folder)
-mask_out_dir = "modal_segmasks"
-
-
-[TRAIN]
-# keep this as is
-model_mode = "training"
-
-# mean pixel for depth images; generally keep this as is
-# should be 128, but for our sim dataset it is actually 119
-mean_pixel = 119
-
-# number of epochs to train for
-epochs = 150
-
-# specify image type
-img_type = "depth"
-
-# learning rate for the model
-learning_rate = 0.01
-
-# image height and image width: remember to use the correct values after resizing
-img_width = 512
-img_height = 384
-# img_nchannels = 1
-
-# base path of the standard format dataset
-base_path = "/nfs/diskstation/projects/dex-net/segmentation/datasets/no_noise_sim_dataset"
-
-# name of the folder to save weights in (within the dataset folder)
-model_weights_folder = "model_weights"
-
-# name of the config file to save (within the dataset folder)
-# note that this will be overwritten between multiple runs so
-# you should name this uniquely for each run.
-save_conf_name = "config_1e-2_e150_resnet50.ini"
-
-[TUNE]
-# keep this as is
-model_mode = "training"
-
-# mean pixel for depth images; generally keep this as is
-mean_pixel = 128
-
-# number of epochs to train for
-epochs = 100
-
-# specify image type
-img_type = "depth"
-
-# learning rate for the model
-learning_rate = 0.0001
-
-# image height and image width: remember to use the correct values after resizing
-img_width = 512
-img_height = 384
-# img_nchannels = 1
-
-# base path of the standard format dataset
-base_path = "/nfs/diskstation/projects/dex-net/segmentation/datasets/real_test_cases_06_07_18/phoxi/images"
-
-# name of the .h5 file to load weights from (within dataset folder)
-base_model_path = "/nfs/diskstation/projects/dex-net/segmentation/datasets/no_noise_sim_dataset/model_weights/no_noise_lr1e-2_e80_clutter_mod/mask_rcnn_clutter_0080.h5"
-
-# name of the folder to save weights in (within the dataset folder)
-model_weights_folder = "model_weights"
-
-# name of the config file to save (within the dataset folder)
-# note that this will be overwritten between multiple runs so
-# you should name this uniquely for each run.
-save_conf_name = "config_no_noise_1e-4_e100_tune_primesense.ini"
-
-
-[BENCHMARK]
-# location of results directory                                                                                                                                                                                                               
-output_dir = "/nfs/diskstation/projects/dex-net/segmentation/results_junk"
-
-# name of results directory                                                                                                                                                                                                                   
-run_name = "junk"
-
-# path to model                                                                                                                                                                                                                               
-model_path = '/nfs/diskstation/projects/dex-net/segmentation/datasets/real_test_cases_06_07_18/primesense_overhead/images/model_weights/primesense_tune20180609T1435/mask_rcnn_primesense_tune_0011.h5'
-
-# path to test dataset (standard format dataset)                                                                                                                                                                                              
-test_dir = "/nfs/diskstation/projects/dex-net/segmentation/datasets/real_test_cases_06_07_18/primesense_overhead/images"
-
-indices_name = 'test'
-
-# Set to True if we want to discard predictions based on if they have bin pixels                                                                                                                                                              
-remove_bin_pixels = True
-
-bin_masks = "segmasks_filled"
-
-# Overlap threshold                                                                                                                                                                                                                           
-overlap_thresh = 0.5
-
-# should always be set to infernece                                                                                                                                                                                                           
-model_mode = "inference"
-
-# name of the config file to save (within the dataset folder)                                                                                                                                                                                 
-# note that this will be overwritten between multiple runs so                                                                                                                                                                                 
-# you should name this uniquely for each run.                                                                                                                                                                                                 
-save_conf_name = "config.ini"
-
-# Visualization settings
-output_pred_vis = True
-show_bbox_pred = True
-show_class_pred = False
-
-output_gt_vis = False
-show_bbox_gt = False
-show_class_gt = False
-
-output_s_bench = True
