--- pcl/config.ini
@@ -1,51 +1,0 @@
-[GENERAL]
-# set to activate a section of the configuration file
-task = "vis"
-
-[VIS]
-test_dir = "/nfs/diskstation/projects/dex-net/segmentation/datasets/real_test_cases_06_07_18/primesense_overhead/images"
-run_dir = "/nfs/diskstation/projects/dex-net/segmentation/final_results/coco_ft_color_primesense"
-indices_name = 'test'
-
-[BENCHMARK]
-# location of results directory                                                                                                                                                                                                               
-output_dir = "/nfs/diskstation/projects/dex-net/segmentation/final_results"
-
-# name of results directory                                                                                                                                                                                                                   
-run_name = "pcl_region_growing_phoxi"
-
-# path to test dataset (standard format dataset)                                                                                                                                                                                              
-test_dir = "/nfs/diskstation/projects/dex-net/segmentation/datasets/real_test_cases_06_07_18/phoxi/images"
-
-# .npy file containing the indices of images to test on for the dataset
-indices_name = 'test'
-
-# Type of PCL detector (either euclidean or region_growing)
-detector_type = 'region_growing'
-
-# Shared options (for both euclidean and region_growing)
-min_cluster_size = 100        # Minimum size of a cluster (default 500)
-max_cluster_size = 1000000    # Maximum size of a cluster (default 1000000)
-
-# Euclidean options
-tolerance = 0.002             # Max distance between closest points in same cluster (default 0.002)
-
-# Region Growing options
-n_neighbors = 5               # Number of neighbors to use for kNN (default 5)
-smoothness = 0.1745           # Max allowed angle in radians between nearby points in same cluster (default 10*pi/180, 0.1745)
-curvature = 0.05              # Max allowed curvature (default 0.05)
-
-# Name of the directory within the test dataset containing bin-vs-not bin masks                                                                                                                                                               
-bin_masks = "segmasks_filled"
-
-# name of the config file to save (within the dataset folder)                                                                                                                                                                                 
-# note that this will be overwritten between multiple runs so                                                                                                                                                                                 
-# you should name this uniquely for each run.                                                                                                                                                                                                 
-save_conf_name = "config.ini"
-
-# Visualization settings
-output_pred_vis = True
-show_bbox_pred = True
-show_class_pred = False
-
-output_s_bench = True
